Scholarship/requirements.txt
apturl==0.5.2
asttokens==2.4.1
attrs==24.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.0.0
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==42.0.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
Flask==3.0.3
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.7.0
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.9.0
Pillow==9.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
reportlab==3.6.8
requests==2.32.3
rsa==4.8
SecretStorage==3.3.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
systemd-python==234
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==2.2.2
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
